Fix for the Reboot Issues on Cypress.
Since the addition of power management, the Cypress driver sometimes gets stuck on reboot. I suspect that this is because the driver has a hard time sometimes turning the touchpad back on due to touchpad firmware timing issues as discussed previously. Using the fixes previously developed to fix rebooting to Linux in the other drivers, this issue can be resolved on Cypress as well.